Move berthForm utils to common/utils/forms
